add some changes to components
