+++ src/components/ui/single_use/Login.tsx
@@ -1,4 +1,5 @@
-import React from 'react';
+'use client'
+import React, { useRef } from 'react';
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
@@ -44,8 +45,31 @@ export function ProfileForm() {
   );
 }
 
+interface FormData {
+  username: string
+  password: string
+}
 
-function Login() {
+import { signIn } from 'next-auth/react';
+interface ILogin {
+  callbackUrl: string
+}
+
+function Login(props: ILogin) {
+  const username = useRef('')
+  const password = useRef('')
+  const { register, handleSubmit, watch, formState: { errors } } = useForm<FormData>();
+  const onSubmit = async(e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault()
+    const result = await signIn('credentials', {
+      username: username.current,
+      password: password.current,
+      callbackUrl: props.callbackUrl 
+    });
+
+  };
+
+
   return (
     <div className="container flex items-center justify-center   ">
       <ResizablePanelGroup
@@ -69,19 +93,20 @@ function Login() {
               <span className="font-bold text-2xl text-primary">Docs</span>
               
             </div>
-            <div className="flex flex-col items-center justify-center flex-1 space-y-4">
+            <form className="flex flex-col items-center justify-center flex-1 space-y-4"  onSubmit={onSubmit}>
             <span className="font-medium text-lg">Welcome To Docs</span>
               <div className="w-full max-w-xs">
                 <label htmlFor="username" className="block text-sm font-medium text-gray-700">
                   User Name or Email
                 </label>
-                <Input id="username" type="text" />
+                <Input id="username" type="text" onChange={e => username.current = e.target.value}/>
               </div>
               <div className="w-full max-w-xs">
                 <label htmlFor="password" className="block text-sm font-medium text-gray-700">
                   Password
                 </label>
-                <Input id="password" type="password" />
+                <Input id="password" type="password" onChange={e => password.current = e.target.value}
+                />
               </div>
               <a
                 href="#"
@@ -89,10 +114,10 @@ function Login() {
               >
                 Forgot Password?
               </a>
-              <Button className="text-mywhite font-semibold rounded-md w-full max-w-xs">
+              <Button className="text-mywhite font-semibold rounded-md w-full max-w-xs" type='submit'>
                 Login
               </Button>
-            </div>
+            </form>
           </div>
         </ResizablePanel>
       </ResizablePanelGroup>
